monitor.py: Don't record '' or ' ' as ShipName
Doing so causes the ShipType to not be used in the UI, so you get a
'link' with just a space for the text.

Any user who purposefully sets their ship's name to a single space
can live with seeing the model name instead.  Yes, I checked, the game
allows it.