Don't run flake8_annotate if no access
This is using github.actor and checking who it is.  If they're not
someone who should have access then use a run with 'exit 1' to fail the
check job.

Then flake8_annotate job needs check_access to have succeeded.

This *will* flag the workflow action as failed if no access, hopefully
it will be obvious why.

Using 'AthanasiusTEST' so it fails for me on this commit/push.