Added timer to inara POST code
This adds an additional wait to the worker thread of the inara code. As
the API key used and all other pertinent data is added to the queue in
call() we can afford here to wait a while between sends without worrying
about going over limits.

I left the original timer fix in place, as it has the added benefit of
ensuring that 30s worth of events are collected _before_ they are
queued. Without that the worker thread would probably end up being
overwhelmed with a bunch of single events.

This commit also breaks the sending logic into multiple functions for
clarity.