Renamed Translations and Locale
Previously at the end of l10n.py, the classes Translation and Locale
were overwritten with singleton instances of their types. This is both
bad practice and makes type checkers want to stab people. This change
should not affect any external code as the names they got for
Translations and Locale were the instances anyway.