Remove last references to, and generation of, systems.p file
This was the last code in eddb.py, so that has now been removed.

This was historically used by the core EDDB plugin to construct a valid
URL for a given station. EDDB has long-since provided alternate URL
formats that negate the need for this.

Plugin authors have been warned not to use this already.

# Conflicts:
#	setup.py