EDMC: Add --trace-on, --trace-all & use monitor.currentdir
In testing `EDMC.py --trace` I ran into issues with NavRoute and
ShipLocker events wanting to open the auxilary files, but it failing.
This turned out to be because that code assumes `monitor.currentdir` is
set and EDMC.py was using a local variable rather than setting and using
it.