companion: Move start_frontier_auth() above login()
The former will be used by the latter, so this makes more sense when
reading the code.

In general I'm going to ensure there are two sections to this Session
class:

1. Frontier Auth code.
2. CAPI query code.

Yes, ideally I'd split it into two classes.  But that kind of
refactoring isn't for *this* branch.  I just want things straight in my
head for the "make CAPI queries run in a thread" changes.