Made plugins UIs frames within frames
Its frames all the way down!

But seriously, the rationale here is that while a plugin may screw with
the column count of itself, it wont screw with anything else. The number
of frames is now silly, however. But that's okay

I still hate tkinter. This took over an hour to get right. But it is
right now, and that's what matters!