CAPI: Avoid '... EDO Settlement?' for 'Update' in space
1. Logged out in space
2. Logged back in
3. Run EDMC
4. Hit Update button, get "Docked but unknown station: EDO Settlement?"

In investigating this I realised that the entire else needed to be "if
docked according to CAPI".

Then it became clear that there was a redundant conditional, as the
enclosing one is already testing:

`capi_response.capi_data['lastStarport']['name'] != monitor.station:`

Making the:

`if last_station != monitor.station`

conditional un-necessary.