github: submodules-update: Port in all the releasehub code & tweak
We want to *check* if there are any changes and only make a branch and add
commits if so.  The upstream code has no option for this, so use it as
a starting point instead.

Specifically this is based on:

https://github.com/releasehub-com/github-action-create-pr-parent-submodule/blob/main/action.yml
aff9d0978a9bbcbc2961d621d5b108c4b46db5e7

* We need 'success' from the *step* to be when there ARE changes, and that
  is in the special github output.
* In order for the whole job *not* to fail in the 'step check_for_changes
  says it failed' case we need it to have `continue-on-error: true`.