requirements-dev.txt: Start listing setuptools explicitly
It turned out that the "fixes semantic_version 2.9.0" version of py2exe
then has an issue with infi.systray because I was still using
`setuptools==60.6.0`, as that's what was latest when I last did anything
to cause its installation in my venv.

So, list it explicitly on the latest version, which works with py2exe from
<https://github.com/py2exe/py2exe/actions/runs/1834917831>, and then
dependabot should prod us to keep it up to date.