monitor: Comment about 'as this User' not actually being true
Both the current `develop` code and this PR's new code share this behaviour.
What the code does is look for windows/processes *in the current Windows
session*, no matter if they're owned by the session user or not.

Neither version will find a process, that would match, if it's running in
another session, i.e. you'd need to switch to it via the windows Login screen.